refactor(commands): 🔧 improve POI naming logic oc:6305 (#216)
* refactor(commands): 🔧 improve POI naming logic

Updated the naming logic for POIs in the `ConvertSuggestedToEcPois` and `ConvertValidatedWaterUgcPoisToEcPois` commands.

In `ConvertSuggestedToEcPois`, the `name` is now constructed using the `ref` and `name` properties if available. In `ConvertValidatedWaterUgcPoisToEcPois`, the `name` is derived from `properties['form']['title']` or defaults to 'Sorgente d'acqua' if not available. Removed an unnecessary line break for better code readability.

* fix(command): 🐛 correct identifier for taxonomy point type

The identifier for `TaxonomyPoiType` was changed from 'water-point' to 'water-monitoring' to ensure accurate classification and prevent potential errors in taxonomy point type retrieval and creation.