refactor(api): 🔄 remove frontend stateful middleware and update CSRF exceptions
This commit removes the `EnsureFrontendRequestsAreStateful` middleware from the API middleware group in `Kernel.php`, as it is no longer necessary for our API endpoints. Additionally, it updates the CSRF exceptions in `VerifyCsrfToken.php` to include various API authentication routes and the `sanctum/csrf-cookie` route to ensure CSRF protection is correctly configured for these paths.

config(cors): 🔧 enable support for credentials

The `supports_credentials` option in the CORS configuration file has been set to `true` to allow requests with credentials, such as cookies or HTTP authentication, to be made to the API.

config(jwt): 🛠️ update JWT configuration for token longevity and claims

The JWT `ttl` (time-to-live) configuration is set to `null`, allowing token expiration to be controlled entirely via environment settings, offering more flexibility in token management. The `exp` claim in the required claims list is now commented out to handle token expiration differently and prevent potential issues with token validation based on expiration.