fix: improve name field handling oc: 6625 (#252)
* fix: improve name field handling oc: 6625

Enhances the retrieval of the name to use a fallback
to the form title when the name is not available.

Relates to oc_6625

* fix: refine display strategy for name field oc: 6625

Improves user interface by refining how the name field is displayed, ensuring it falls back on the form title if the specific name is not available.

Relates to oc_6625

* feat: add POI creation helper text oc: 6625

Adds a user guidance section to POI creation/edit process, ensuring
that users can follow steps intuitively. Includes conditional display
logic to aid users until mandatory fields are completed and updates
localization files with new instructions.

Relates to oc_6625

* Fix styling

* fix: make name field readonly oc: 6625

Updates the name field to be readonly in the common fields trait
to ensure it cannot be modified through unintended means.

Relates to oc_6625

* style(nova): 💄 improve code readability and consistency

- Utilize `data_get` for cleaner property retrieval in `Text::make('Name', ...)`
- Adjust spacing around concatenation operators in `->displayUsing(...)` for consistency
- Remove unnecessary space in closure parameter within `validatedStatusOptions` function for cleaner syntax

* Fix styling

* Updated submodule wm-package

* refactor(Nova): ♻️ optimize helper text and status display

- Improved the structure of helper text in `UgcPoi.php` by using the HEREDOC syntax for better readability and maintainability.
- Refactored the status display logic in `UgcCommonFieldsTrait.php` to maintain consistent spacing and improve code clarity.
- Ensured all string concatenations are consistent and follow the same coding standards for better code quality.

* fix: improve name field handling oc: 6625

Enhances the retrieval of the name to use a fallback
to the form title when the name is not available.

Relates to oc_6625

* fix: refine display strategy for name field oc: 6625

Improves user interface by refining how the name field is displayed, ensuring it falls back on the form title if the specific name is not available.

Relates to oc_6625

* feat: add POI creation helper text oc: 6625

Adds a user guidance section to POI creation/edit process, ensuring
that users can follow steps intuitively. Includes conditional display
logic to aid users until mandatory fields are completed and updates
localization files with new instructions.

Relates to oc_6625

* Fix styling

* fix: make name field readonly oc: 6625

Updates the name field to be readonly in the common fields trait
to ensure it cannot be modified through unintended means.

Relates to oc_6625

* style(nova): 💄 improve code readability and consistency

- Utilize `data_get` for cleaner property retrieval in `Text::make('Name', ...)`
- Adjust spacing around concatenation operators in `->displayUsing(...)` for consistency
- Remove unnecessary space in closure parameter within `validatedStatusOptions` function for cleaner syntax

* Fix styling

* Updated submodule wm-package

* refactor(Nova): ♻️ optimize helper text and status display

- Improved the structure of helper text in `UgcPoi.php` by using the HEREDOC syntax for better readability and maintainability.
- Refactored the status display logic in `UgcCommonFieldsTrait.php` to maintain consistent spacing and improve code clarity.
- Ensured all string concatenations are consistent and follow the same coding standards for better code quality.

* Fix styling

* refactor: relocate UGC creation helper oc: 6625

Moves POI creation instructions from the UgcPoi class to
the UgcCommonFieldsTrait for improved code organization and reuse. Updates
language strings for consistency.

Relates to oc_6625

* refactor(UgcCommonFieldsTrait): ♻️ extract creation helper logic to separate method

- Moved the creation helper HTML and logic into a dedicated `getCreationHelperHeading` method.
- Improved code readability and modularity by encapsulating the helper text generation.
- Adjusted `getCommonFields` to utilize the new method for creating the helper heading.
- Ensured that the visibility logic for the helper remains consistent with prior implementation.

---------

Co-authored-by: RubensGarofalo <RubensGarofalo@users.noreply.github.com>
Co-authored-by: peppedeka <peppedeka@gmail.com>
Co-authored-by: peppedeka <peppedeka@users.noreply.github.com>